moved merge simul tmp hts to fix circular import issue
